Re-adds page-level props & ensure backwards compatability
Using `upsertBlockProperty` gives us the desired behaviour:

1. When a property doesn't exist, we create it and set it to the
   provided value.
2. When a property does exist, but it doesn't match our value, we set it
   to match our value.
3. When a property does exist and matches our value, we don't do
   anything.

The previous function, `updateBlock`, has unexpected (or non-obvious
behaviour, anyway) where it clears a blocks content, including the
properties, if the text is empty (""). That was clearing the page-level
properties instead of updating them. Not ideal, really.

`upsertBlockProperty`, on the other hand, seems to work nicely.